[Erlang] Fix more emergency bailouts
This commit adds (some) missing emergency bailouts to pop off from
scopes not after a clause's end.

Effects:
 - groups
 - lists
 - maps
 - record fields
 - tuples

It also adds the `illegal-stray` context to scope the last closing
parentheses/bracket/brace `invalid.illegal` in order to indicate an
syntax error before the end of the clause. The idea behind is any
nested unclosed context to with the `.`.