[Erlang] Fix fun call identifier
The `variable.function` scope is generally used to scope an function
identifier in all default syntaxes. A normal function identifier in
Erlang is an atom. Means a normal function name must start with lower
case or underscore.

  my_func_name()

On the other hand, Erlang supports the definition of closures via fun
keyword, which binds a "function" to a real variable.

  // define the function
  MyLocalFunc = fun() ... end,

  // call the funciton
  MyLocalFunc(),

Scoping both, an ordinary function identifier and a closure as
`variable.function` is therefore confusing. As we already support
constants or macros to be used as identifiers

  %MY_FUNC_MACRO()

it is logical to scope `MyLocalFunc` as `variable.other` because it
is a real variable, which holds a function.