[ShellScript] Control Structure Improvements (#1863)
* [ShellScript] Rename control flow keywords

This commit...

1. renames all control flow keywords to comply with
   https://github.com/sublimehq/Packages/issues/1228

2. renames the "case" contexts to comply with sematics of
   https://www.gnu.org/software/bash/manual/bashref.html

   Each case branch consists of a `pattern` and a `clause`, which is
   executed if the `pattern` matches the `word`.

* [ShellScript] Fix case context stacking

This commit fixes a design issue of the case context, which could probably cause the parser to hit a sanity limit if a file with too many case clauses was loaded.

Instead of chaining each case clause via `set` keyword until `esac` is matched, a common `case-body` context is introduced from which each pattern/command-list pair is pushed. This results in at most two contexts to be pushed to the scope stack no matter how many clauses the case statement contains.

* [ShellScript] Fix keyword boundary lookahead

This commit adds a `keyword_boundary_char` and reuses it in the `keyword_boundary_end` variable.

Notes:

- The `_` is removed as it is part of \w
- The negative lookahead `(?!=)` is merged with `[^\w_-]`
  with the assumption that a keyword is not matched as such
  if followed by a `=` or `-`.
- The `$` is removed because `[^-=\w]` seems sufficient to
  match a word boundary

As a result the new variables can be used for the keywords in the recently refactored `case` statement to harden keyword matching.

* [ShellScript] Rename 'keyword_boundary_end' to 'keyword_break'

Inspired by other syntax definitions such as JavaScript this commit renames the 'keyword_boundary_end' to 'keyword_break'.

This change has no functional effects.

* [ShellScript] Fix wrong indention

This commit fixes a lint issue with an under-indented context list by turning it into an array, which was chosen as the number of contexts and the length of the names is short enough to look well.

* [ShellScript] Address requested review changes

1. The `keyword_break` is turned into a negative lookahead to make
   it work properly at the eof without breaking its function in
   other situations.

   Note: Adding `$` would have broken the syntax definition.

2. The keyword `elif` is scoped with the standardized `elseif`
   sub scope.

3. The keyword `done` is scoped as `... loop.end` as it finalizes
   a for loop. This scope now matches the conditional statements,
   which use `keyword.control.conditional.end` already.

* [ShellScript] Fix 'esac' highlighting in incomplete case statement

This commit fixes an issue which caused the the `esac` keyword not to be matched if the `in` keyword is missing.

During fixing the first issue it turned out the `keyword_break` variable can be used in the `case-end-ahead` context as negative lookaheads work well in positive ones.

* [ShellScript] Highlight stray `esac` as keyword like stray `fi`.

If the leading `case` keyword is missing or incomplete, the finalizing `esac` should be scoped as keyword as well.

This commit does not scope it as illegal as it is not yet implemented for `fi` as well. The syntax does not handle many illegals at all, so this commit does not intend to change that.

* [ShellScript] Fix meta.conditional.case boundaries

This commit ensures to add the meta scope to the `esac` keyword only if it finalizes a valid `case` statement.

* [ShellScript] Fix then keyword_break

This commit applies the {{keyword_break}} lookahead to the `then`
keyword in order to not scope `then-` and friends as such.

* [ShellScript] Fix highlighting of ! after logical operator

Fixes #1868
Fixes #1871

* [ShellScript] Simplify cmd-post context

This commit simplifies the `cmd-post` context by:

1. removing the `meta.post-cmd.shell` which seems to have no
   function and is not part of any completion or test
2. removing the initial `set` command which is no longer needed
   due to removing the meta scope
3. removing some once-used variables which seem to create more
   confusion then improve readability
4. removing capture groups to improve parsing performance as ST
   is known to win about 12-15% by not using `caputure:`

* [ShellScript] Fix `!` operator

This commit fixes a regression with not highlighting `!` as operator
at the beginning of a line.

According to http://tldp.org/LDP/abs/html/special-chars.html#NOTREF

`!` followed by a space => keyword
`!` not followed by a space => triggers shell history function

This rule is applied to ShellScript now.

* [ShellScript] Fix `!` at end of file

This commit applies a requested change from review which fixes
the issue of not matching `!` as keyword.operator.logical, if
it is followed by the end of file.

* [ShellScript] Remove obsolete lookahead

The lookahead in the second `!` match is not needed.

Also add some tests to ensure we don't break with line continuation.

* [ShellScript] Add support for shell history references

According to @FichteFoll's comment and a description at (1)
the shell supports `!!`, `!-<n>` and `!<n>` to access commands
from the shell's history.

This commit scopes these tokens as `variable.language`.

(1):  https://www.digitalocean.com/community/tutorials/how-to-use-bash-history-commands-and-expansions-on-a-linux-vps