[Perl] Add array item access (#1953)
* [Perl] Add array item access

This commit adds support to scope the item access of an array.

Examples:
  $array[0][2]
  $dict{'key1'}{'key2'}

This change is required to properly highlight those expressions in
string interpolations. Otherwise the array index would not be part of
the interpolation.

It turned out an array is initialized with the following expression:

  $array = [ item1, item2, ...];

This commit adds the "brackets" contexts with rules to scope the
opening and closing brackets.

Note:

  1. The `groups-nested` is not added everywhere to reduce the risk of
     interfering with HEREDOCs. It is assumed no one decides to define
     a HEREDOC within an array item access expression.
  2. The `groups-nested` was not yet used and contained a therefore
     unrevealed issue, which is fixed as well. It used the wrong match
     patterns. `{` instead of `(`.

* [Perl] Add referenced iten access

This commit adds the meta.item-access to expressions like

  $array->[0]
          ^^^ meta.item-access

to create a consistent scoping.