[Git] Add .mailmap highlighting (#1876)
* [Git] Add .mailmap highlighting

It's very minimal, but names and emails are indexed

* [Git] Mailmap non-word names and incomplete emails

* [Git] Rename Mailmap files

* [Git] Simplify .mailmap (#3)

This commit replaces the external reference to `Git Common.sublime-syntax#email` by an extended version of the existing lazy email pattern, defined in the syntax itself as most of the email parts are already there.

Furthermore it removes matching `[ \t]*` in front of the "proper name" as it does not have any effect.

* [Git] Fix punctuation in email references (#4)

This commit makes sure all dots in the domain name of an email address are scoped as `punctuation.separator.email`. It furthermore enables the syntax to scope the name and domain part differently if desired.

* [Git Formats] Optimize MailMap pattern

This commit removes the capture group from the Git Mailmap syntax.
This is pure optimization. ST's sregex engine is known to be 10-13%
faster by avoiding `capture`.

Benchmark:
   
   - ST 3207
   - Win10

   A file with 100k lines of: 

      Some Dude <some@dude.xx>

Result:

   before: 490ms
   after: 440ms  
   faster by: 10%

* [Git] Rescope email domain dot