[Perl] Add unquoted hash keys highlighting
In order to access an item of a hash (dictionary), the caller can use
an unquoted string as key.

In the other way round: The `item_key` in `$the_hash{item_key}` is a simple unquoted string.

Prior to this commit, the key was not scoped at all and would so as function/method call in the future.