[Perl] Add string interpolation step 3
This commit adds a `variable-interpolation` context, which can be
included wherever interpolation support is needed.

It removes the last scope on stack, which should be a subscope of
`string` and adds the `meta.interpolation.perl`.

This commit adds interpolation support to:

 - double quoted strings
 - backtick quoted strings
 - format strings

Note: Perl does not evaluate variables in single quoted strings.