[Perl] Restrict whitespace surrounding namespace accessors
A fully qualified identifier must not contain whitespace. Hence the
accessor `::` must not be surrounded by them as well.

It was supported up to this point in order to improve the writing
experience by not breaking identifiers during writing.

This will cause conflicts with parsing/scoping fully qualified
identifiers in future changes an therefore needs to be restricted.

The test cases containing whitespace surrounded accessors are removed
for the moment as properly handling such situations is part of a future
change.