[Perl] Tweak namespace scopes
This commit renames `support.class.perl` to `variable.namespace.perl`
as more general identifier for the path parts of a qualifier.

The top-level scope `support.` is not sufficient as it is meant for use
with built-in entities only. Any namespace can be user defined though.

The `variable.namespace` was choosen as counterpart of
`entity.name.namespace` which is used to define a namespace.