[Perl] Fix interpolated string termination
This commit addresses interpolated strings and quoted like functions
not to be correctly popped off from, if the interpolated string ends
with a variable punctuation.

Examples:

  a) "$repl$"
           ^^ no variable!

  b) s/pattern/$repl$/g;
                    ^^ no variable!

Example (a) results in a syntax error as the Perl interpreter doesn't
know whether to handle the `%"` as variable or not. This commit does
not introduce an `illegal` scope for it, though. It just ensures not
to break the string boundaries.

Example (b) shows quoted like functions, which Perl tokenizes by the
first character after the function identifier (here `/`) first before
it starts parsing the strings. The delimiters are therefore matched
with higher priority.

This commit adapts this behavior by

1) consuming the variable punctuation `[$@%&*]#?` in front of the
   delimiter (or closing bracket).

   We need it because, $/, $), $], etc. are valid built-in variables.

2) modifying the `variables-interpolation` context in order to make
   sure to pop off from a `meta.interpolation` after each variable.
   Otherwise 1) wouldn't work.

   In order to prevent sophisticated (and error prone) lookaheads for
   each ordinary variable, the existing `qualified-variables` and
   `unqualified-variables` contexts are merged for that purpose.

   The resulting context pops off, while the original ones don't.