[Perl] Tweak function vs. constant matching
Single words within expressions are most likely to be constants.
They can be functions if defined by an import (use) or sub before,
but we can't distinguish that.

Example:

    if (constant)
        ^^^^^^^^  constant or function possible.

Uppercase only identifiers are already scoped as constants, if they
don't look like a namespace or function call.

This commit adds the `constant-identifier` context to scope all
identifiers as constants, which were not otherwise matched.

This is only guesswork but the only chance we have to hopefully scope
everything correctly in most situations.