[Java] Fix lowercase inherit interface identifier
Fixes #1858

This commit is a local hotfix for issue #1858. If the identifier of an
inherit class starts with a lowercase character, the rest of the
class definition becomes broken. Therefore the methods aren't properly
scoped and not part of the index.

Supporting lowercase here, would mean to support lowercase class
identifiers in general to be complete, but this seems to be a bit more
complicated.