[Ruby] Fix regexp patterns beginning with braces
Fixes #1515

This commit modifies the lookahead of the `try-regex` pattern in order
to allow the pattern to start with `{`.

The `=` is the only character which causes Ruby to not interpret a
`/.../` expression as regexp-match.

All the other ones (`*`, `+`, `?`) trigger regexp related errors.

The first `{` is interpreted as normal literal character.