[HTML] Split class/id attribute values
This commit attempts to support a fix for
https://github.com/SublimeTextIssues/Core/issues/2960

Up to this commit no scope exists, which represents a single item of
that list. It's therefore impossible to use it in a "Goto Definition"
scenario.

This commit therefore adds a `tag-attribute-value-separator` context
in order to split the `meta.class-name` and `meta.toc-list.id` tokens.

The `class` and `id` attributes' quoted values may represent a list of
classes/ids to describe the html tag.

Note:

The decision was made to add the `tag-attribute-value-separator` which
clears the `meta.class-name` and `meta.toc-list.id` because the
required regexp patterns are much simpler and more safe than for an
otherwise required `tag-attribute-value-item` context, which
explicitly matches the single classes/ids.