[Git Formats] Embed commit message into git log
This commit uses the `embed` command instead of `push with_prototype`
to include the Git Commit.sublime-syntax into Git Log.sublime-syntax

It is a performance tweak (about 5% faster).

Note:
  
This commit contains a workaround for 

  https://github.com/SublimeTextIssues/Core/issues/2395

as the `Git Commit.sublime-syntax` "sets away from main", which causes
the `escape` command to be ignored with all current builds of ST3.

The second hunk can be reverted as soon as the core issue is fixed.