[JavaScript] Update keyword scopes
This commit modifies control structure keyword scope names to make
them comply with the results of the discussion in issue #1228.

The goal is to make all default syntaxes use the same precise scopes
for related keywords.

Notes:

The following keywords haven't been discussed yet:

1) `await`    -> `keyword.control.flow.await`
2) `debugger` -> `keyword.control.flow.debugger`
3) `with`     -> `keyword.control.flow.with`

The `with` might need to use another scope though, as it has a slightly
different meaning.