[Diff] Add syntax specific setting to enforce tabs
This adds a `Diff.sublime-settings` file that reinforces that the
default  setting for `translate_tabs_to_spaces` in Diff files should be
`false` unless the user specifically overrides it.

This is due to the fact that if you enable that setting in your overall
preferences, diff output that Sublime generates may convert any tabs in
the file to spaces, and arguably the diff output should always show the
input files in as true a representation as possible.