Force tab translation off during diff display
After discussion in the original PR this change proposes that the
default setting for trimming trailing whites pace should be disabled in
diff/patch files instead of forcing tab translation off; saving changes
to a patch (even inadvertently) shouldn't potentially cause it to no
longer apply (or apply incorrectly).

In its place, this instead temporarily forces tab translation off while
writing the diff output to the view/panel to ensure that the output uses
the correct characters.

This ensures that the headers get written correctly, since the unified
diff "spec" indicates that the names of the files and their modification
dates in the header should be separated by tab characters and external
tools may rely on that.