[Perl] Improve numeric highlighting (#2032)
* [Perl] Fix numeric constant boundaries

This commit fixes two issues with numeric constants.

1) Some of the constants were not scoped correctly if directly followed
   by a `+` operator, which was caused by the `+` in the `(?![\.\w+])`
   lookahead, which was probably added by accident.

   Investigations revealed Perl correctly detecting all numbers even
   if followed by `.`. It then just handles that dot as concat operator.

   Example:

       `print 0x1.0x2` will print out 12

2) It's barely possible to distinguish a negative number from a
   subtraction operation. Therefore the leading number sign matching
   `[-+]?` was removed from unquoted constants.

   All `-` or `+` operators in front of or after numbers are now
   scoped as `keyword.operator` and do no longer break highlighting.

* [Perl] Add scope for octal numbers

This commit scopes integers starting with `0` and containing digits
from 0 to 7 as `constant.numeric.integer.octal.perl`.

* [Perl] Fix incomplete floating points

Perl accepts all kinds of incomplete floating point numbers.

This commit therefore scopes all of them properly.

This commit doesn't split the `match` operations for performance
reasons. The current implementation was benchmarked the fastest one
from various alternatives tested against 10k lines of quoted and
unquoted floating point numbers.

Note:
  Need to make sure not to brake the range operator `..`.

* [Perl] Add scope for decimal separator

Languages like Go or Python scope the decimal separator already.
With the changes made in the previous commit doing so is too easy to
ignore it.

* [Perl] Add scope for bin/hex/oct punctuation

Languages like D scope the leading `0b` or `0x` of binary/hexadecimal
constants as `punctuation.definition.numeric`.

This commit adds the according rules to do so for Perl as well for
consistency reasons with other languages.

* [Perl] Fix unsupported quoted octal numbers

Quoted numbers are interpreted as decimal (float or int) only.
Octal numbers as binary and hexadecimal must be unquoted.

Example:

  print "030" + "10";    # prints 40

* [Perl] Merge the constant-number contexts

As the both removed contexts are not used separately at any point they
are merged into the `constants-numbers` context.

The rules are sorted by number bases: bin/oct/hex/dec

* [Perl] Fix bin/hex/float boundaries

Merging the numbers contexts revealed some missing boundary checks.
Even though they are considered not needed they are added for safety
and consistency reasons to prevent possible edge cases.

* [Perl] Fix floats being treated as octal

According to the test cases of ModernPerl 01.0 is a decimal float, but
was treated as octal due to the current order of the matches.

This commit resorts the rules to fix it.

* [Perl] Fix case of bin/hex/float punctuations

Case doesn't matter, when parsing 0x, 0b, or exponent e. So 0X, 0B or E
are valid as well.

* [Perl] Fix numbers with underscore

Perl allows all digits in a number to be replaced by `_`.