[Ruby] Update string interpolation (#2054)
* [Ruby] Fix interpolated require

Fixes #2000

This commit causes the `require` context to pop off by `}` in addition
to the end of a line in order to make it work properly within string
interpolations.

It also removes some unnecessary captures.

* [Ruby] Add meta.string.ruby

This commit prepares for the new `meta.interpolation` support.

* [Ruby] Add meta.constant.ruby

This commit prepares for the new `meta.interpolation` support.

The quoted constants (`:"constant"`) support interpolation and need to
be handled like quoted interpolated strings. To be able to remove the
`constant` scope, a corresponding meta is added.

* [Ruby] Update constant/string interpolation

This commit makes sure to remove the `string.` or `constant.` scopes
from interpolated expressions and variables as proposed in the latest
scope naming guidelines.

* [Ruby] Fix embedded source scopes in HEREDOC

SQL, CSS, JS and RUBY are `source.` not `text.`

* [Ruby] Add string interpolation to HEREDOCs

This commit applies the minimal set of changes in order to properly
support the `meta.interpolation` vs. `meta.string` scopes with HEREDOCs.

1) Embedded syntaxes (SQL, HTML, ...) are handled as interpolations,
   which means the `string.` scope is removed while keeping
   `meta.string` and the embedded `source.`/`text.` only.

2) The `interpolated-...` context is moved up to make variable 
   interpolation work "a little bit".

Notes:

1) To make interpolation work properly `with_prototype` was needed to
   inject `interpolated-...` into any context of the embedded syntax.

2) Not sure whether escapes are really supported/useful when used with
   embedded syntaxes.

3) Ruby 2.3 introduced single quoted HEREDOCs without interpolation
   support, which is not yet included in this syntax definition.

4) While HEREDOCs work pretty much like in Bash/Perl/PHP all 
   implementations are very much different from how the single tokens
   are scoped. We may need an RFC discussion to come up with a standard.

* [Ruby] Add support for all the different HEREDOC styles

Based on the existing structure and scope naming this commit adds
support for all the different HEREDOC styles:

* indented/unindented end tag
  - unindented end tag: `<<TAG`
  - indented end tag: `<<-TAG` or `<<~TAG`
* interpolated/plain content
  - interpolated: unquoted, double-quoted, backtick-quoted TAG name
  - plain: single quoted tag name

SEE: https://ruby-doc.org/core-2.5.0/doc/syntax/literals_rdoc.html

Notes:

1) Support for the existing embedded syntaxes keeps unchanged.
   Interpolation might not work sufficient enough for those syntaxes
   which push/set into contexts though, which was an issue before and
   is not fixed by this commit.
2) Support for embedded shell syntax is enabled again as it seems to
   work well with the current Bash.sublime-syntax

* [Ruby] Remove capture group from HEREDOC patterns

The outer most capture group is not used and therefore removed.

* [Ruby] Use variables for HEREDOC delimiters

This commit intends to

1. increase readability
2. limit the syntax prefix to uppercase as the syntax name
   Therefore `(\w+_)` is replaced by `([[:upper]_]_)`

* [Ruby] Fix regression

The `expressions` context should be included in HEREDOCs with embedded
Ruby only.

* [Ruby] Update TODO list

This commit removes single quoted uninterpolated HEREDOCS from the TODO as it is fixed by this PR/branch.

* [Ruby] Cleanup meta.string scopes of regexp patterns

* [Ruby] Fix regexp patterns beginning with braces

Fixes #1515

This commit modifies the lookahead of the `try-regex` pattern in order
to allow the pattern to start with `{`.

The `=` is the only character which causes Ruby to not interpret a
`/.../` expression as regexp-match.

All the other ones (`*`, `+`, `?`) trigger regexp related errors.

The first `{` is interpreted as normal literal character.