[Python] Simplify integer scope names
This commit renames the integer scopes to comply with the scope naming
guidelines by removing the `long` sub scope.

The `long` is already expressed by the `storage.type` suffix. Except
the `L` suffix indicates the data size it is nearly impossible to
guess the data type of a numeric constant in general.
Therefore it seems better to have common scopes for all numerics.