[Ruby] Scope alias/undef as keyword
The keyword `alias` is somehow like `#define` in C. It is used to
declare an alias at compile time. The `undef` does the opposite.

As there are two directives for defining and undefining an alias, the
`storage.type` scope wouldn't be appropriate here.

An alternative would be `keyword.control.directive` which is used for 
preprocessor like keywords in Erlang.