[Erlang] Fix quoted identifier lookahead
Fixes https://github.com/sublimehq/Packages/pull/1878#issuecomment-546978653

This commit improves the quoted identifier pattern as the old one
caused the namespace lookahead to be triggered in the wrong place.

The following token was matched by the (?={{ident}}\s*:[^:=]) pattern
in the namespace context by accident:

   '$get_stacktrace') catch exit:'$get_stacktrace':

The new pattern makes sure to stop matching at the first unescaped
single quote. Any number of escaped backslashes is allowed in front of
it.