[Python] Improve numbers (#2136)
* [Python] Improve float and complex numbers

This commit

1. applies scope naming guidelines to imaginary numbers.
2. optimizes float/imaginary number parsing by slightly smaller rules,
   which improves performance by about 14%

Note:
   The basic patterns were borrowed from Perl.

Benchmark:

  10k of the following lines were used to compare the performance of
  the old and the new solution:

  .1 .1e1 .1e-1 .1e+1 1.1 1.1e1 1.1e-1 1.1e+1 1.e1 1.e-1 1.e+1 1.1e1 1e-1 1e+1

* [Python] Simplify integer scope names

This commit renames the integer scopes to comply with the scope naming
guidelines by removing the `long` sub scope.

The `long` is already expressed by the `storage.type` suffix. Except
the `L` suffix indicates the data size it is nearly impossible to
guess the data type of a numeric constant in general.
Therefore it seems better to have common scopes for all numerics.

* [Python] Slightly performance tweak

Found the modified `digit` pattern in PR #2123 for Go. It is slightly
faster than the current one.

* [Python] Cleanups

This commit...

1. renames the `digit` to `digits` as the latter one is more appropriate.
2. sort character classes to always start with lower case letters
3. add the `decimal` sub scope to imaginary numbers to comply with the
   scope name of R, which distinguishes between decimal and hexadecimal
   imaginary numbers. They are implicitly considered floats.

* [Python] Remove .complex sub scope

* [Python] Simplify numeric suffix scope

This commit applies a review request.

Suffixes of numeric literals don't need to be too specific as ...

1. they apply to single characters only
2. should all be colored in the same way, especially if some literals
   have multiple ones.