[YAML] Improve numbers (#2171)
* [YAML] Improve numbers

This commit adds some capturing groups to the existing numeric literal
match patterns in order to distinguish the types (int/float/...) and
to add scopes for the separators.

* [YAML] Rescope the indention indicator

The indention indicator is an ordinary integer literal and thus should
be scoped as such. Like the file-descriptor in ShellScript, the special
meaning is then appended as sub-scope to the ordinary numeric scope.