[Ruby] Scope super as function
In Ruby, `super` behaves like a magic function that resolves to the
appropriate function. Compared with keywords like `self` or `this` in
other languages, which are scoped as `variable.language`,
`support.function.builtin` seems to make most sense here and complies
with Python, even though the meaning is a little bit different here.