[CSS] Completions: Complete values with semicolon
Fixes #2076

This commit creates a consistent behavior when either completing
property names or typing `:` after a complete property name by making
sure a property name is always followed by `: ;` with the caret positioned directly in front of the semicolon.

Up to this commit a `;` is added directly after a typed `:` without
any spaces in between and without respect whether a semicolon already
exists after it.

On the other hand no `;` is added after a property name, which was
inserted by the `css_completions.py`.