[CSS] Keep cursor in front of semicolon after value completion
This commit is to further improve consistency with property completions.
Up to this commit all completions cause the caret to be placed in
front of the semicolon, except the case a property-value is completed
while no semicolon is followed.

Behavior up to this commit:

    // before

    color: |

    // committed completion

    color: black;|

Behavior introduced by this commit:

    // before

    color: |

    // committed completion

    color: black|;

Note:
  The pipe is the caret position.