[CSS] Tweaks for the keymap
This commit adds some comments to describe the purpose of a key binding
and adds some checks to avoid edge cases.

The colon snippets are fired only if the preceding character is a word
character, which indicates the end of a property name.

The Add Line macro is fired only if the caret is not between two
semicolons, but in front of the only one which terminates a value.