[Ruby] Improve keyword scopes (#2156)
* [Ruby] Fix 'class' scope name

This commit applies `storage.type.class` to the `class` type keyword.

* [Ruby] Fix 'module' scope name

This commit applies `storage.type.module` to the `module` type keyword.

* [Ruby] Fix 'def' scope name

The keyword `def` declares a method/function and is therefore scoped
as `storage.type.function`.

* [Ruby] Expand keywords scopes

This commit applies #1228.

Note:

The keywords `begin`, `end` and `do` are used in several situations to
open or close a block of code or entering a context. They are therefore
scoped as `keyword.control.block` as synonym for `meta.block` and
`punctuation.section.block.[begin|end]`.

* [Rails] Fix test case

This commit applies required changes caused by changes in Ruby.

* [Ruby] Scope alias/undef as keyword

The keyword `alias` is somehow like `#define` in C. It is used to
declare an alias at compile time. The `undef` does the opposite.

As there are two directives for defining and undefining an alias, the
`storage.type` scope wouldn't be appropriate here.

An alternative would be `keyword.control.directive` which is used for 
preprocessor like keywords in Erlang.

* [Ruby] tweak declaration keyword scopes

* [Ruby] Merge fail,raise,throw scopes

* [Ruby] Scope super as function

In Ruby, `super` behaves like a magic function that resolves to the
appropriate function. Compared with keywords like `self` or `this` in
other languages, which are scoped as `variable.language`,
`support.function.builtin` seems to make most sense here and complies
with Python, even though the meaning is a little bit different here.

* [Ruby] Allow regexp like super arguments

As super is treated as function, it should accept regexp like arguments.