[Java] Add receiver parameter to method declarations
The first parameter of a class method may be a so called receiver 
parameter, which denotes the class instance the method is associated
with. It is not a normal formal parameter its identifier may be 
qualified with a class name.

see: https://docs.oracle.com/javase/specs/jls/se13/html/jls-8.html#jls-ReceiverParameter