[Java] Split primitive-array and object-array contexts
This commit splits the object-and-array-types to be able to handle
the different aspects of scoping primitive types in parameter lists
and generic-arguments lists in a more precise way.

This commit makes sure to correctly handle annotations especially in
generic type arguments by using the new branch feature. It's the only
way to properly support annotations with fully qualified identifiers
and /or parameters in generic type arguments.