[CSS] Fix auto-pair quoted values
By automatically adding a semicolon to the end of line, when committing
completions, the normal situation for starting values is the caret to be
located in front of it (semicolon).

Example:

   content: |;

ST's default key bindings don't auto-pair a single or double quote if
`auto_match_enabled` is `true`, because the `;` is not part of the
`following_text` context.

Pressing `"` therefore results in:

   content: "|;

This commit fixes this behavior when editing CSS files so the result is:

   content: "|";

It also helps in situations of:

   content: func(|,);

to make them expand to:

   content: func("|",);