[ShellScript] Get rid of all the -bt contexts
Fixes #1781

This commit makes use of the `embed` feature to remove duplicated
`-bt` contexts, which were added to support backtick-expansions.

Every kind of command is now supported within backtick-expansions even
for loops.

With regards to the benchmark scenario below this solution is about 20%
slower than the old one, but helps to keep the code clean and
maintainable. The overall impact on a script depends on the amount of
backtick-expressions being used.

Benchmark:

  10k lines of: `echo hello world`