[ShellScript] Tweak vardef/funcdef contexts
This commit removes the leading `\s*` and related capture groups from
the rules of the vardef and funcdef context.

The goal is to match those declaration keywords the same way as control
keywords. See the differences of `function` and `coproc` which have been
removed for instance.

A positive side effect is a 10% improved overall parsing performance of
the syntax test file.