[HTML] Fix </style> tag matching
This commit works around a core issue which prevents lookaheads within
lookaheads from being applied correctly, which caused the escape of the
embedded CSS not to be applied due to the recent changes.

The same fix is applied to embedded scripts as well, even more. Make
sure to leave a script tag only if a valid </script> tag was matched.

Note:

The issue triggers only if the HTML syntax is used by an embedding
syntax like HTML Jinja, which makes it impossible to add a test case.