[Markdown] Separate bibliography reference links
Fixes #2360

Reference definitions need to be separated by line to correctly display
them in the Goto Symbol Overlay.

Example:

[foo] https://foo.com
[bar] https://bar.com
[baz] https://baz.com

The newline `\n` is no longer included into the `meta.link.reference`
scope in order to separate the lines from each other.

As `\s*` includes the `\n` the `link-title` context is modified
accordingly.

The former `after-link-title` and `link-ref-def` contexts are merged
into a `link-ref-def-expect-end`