[Erlang] Use storage.type.record
The #record{field=value} term represents an instantiation of a record,
which was declared by `-record(record, {field=value})` statement. Thus
the #record identifier can be interpreted as storage.type.

This commit uses `storage.type.record` instead of `storage.type.struct`
as this scope name is used for the meta scopes as well.