[ShellScript] Fix context pop before comments
All single line contexts need to be popped off the stack before comments
start in order to prevent line bleeding. Means, comments must not cause
line boundaries to be ignored causing command argument contexts to
continue on the next line.

To achieve that the `comment` context is removed from `prototype`.