[ShellScript] Workaround core issue 2482
This commit creates a dedicated `statements` context which can be
embedded or included without adding a second `source.shell` scope.

This was the case with the `embed` statement to support backtick command
expansions before this commit.

It also makes the workaround for 1062 stand out, clearly.
The `meta_include_prototype: false` is obsolete then.