[Java] Move prototype fix to server pages
The prototype is not applied when `- include: scope:source.java` only.
It works as expected when using `push/set/embed: scope:source.java`.

Hence it is the responsibility of the JSP syntax to explicitly include
the prototype context if required. We don't want a general duplicate in
Java's `main` in general.