[Regex Replace] Add syntax (#2214)
This commit adds the Regex Replace.sublime-syntax to highlight the
content of the replace input widget if the regexp mode is enabled.

The knowledge about the Regex Replace Widget.sublime-setting was
published by @keith-hall at Discord.

The syntax definition is sourced from PackageDev with some fixes and
scope name modifications being applied. Scoping of conditional
expressions and groups was removed as ST doesn't seem to support that.

Both file name and the `name` attribute use `Regex` as the widget
setting is named like that. The scope name on the other hand uses
`regexp` as this is what the general `Regular Expressions` syntax
defines.