[HTML] Consolidate tag/attribute name matching (#2239)
* [HTML] Consolidate tag/attribute name matching

This commit ...

1. consolidates the way the case-insensitive flag `(?i)` is applied to
   match rules. All remaining rules are modified to apply it as local
   flag for strings only.

   Notes:
   a) The `{{javascript_mime_type}}` is case-insensitive already.
   b) It is removed from lookaheads with non-word characters only.
   c) All tag attributes are matched case-insensitive now.

2. removes the leading `\b` in some remaining tag and attribute names
   as the way the syntax works already ensures any match to start at
   an appropriate position (not in the middle of a word).

3. replaces the trailing `\b` in some remaining tag and attribute names
   by `{{tag_name_break}}` or `{{attribute_name_break}}` to make sure
   to properly match the name boundaries according to the whatwg specs
   under all circumstances.

4. consolidates the `</script>` and `</style> closing tag match rules
   to behave like any other recently updated tags. The `>` can be at
   the next line now.

* [HTML] Fix </style> tag matching

This commit works around a core issue which prevents lookaheads within
lookaheads from being applied correctly, which caused the escape of the
embedded CSS not to be applied due to the recent changes.

The same fix is applied to embedded scripts as well, even more. Make
sure to leave a script tag only if a valid </script> tag was matched.

Note:

The issue triggers only if the HTML syntax is used by an embedding
syntax like HTML Jinja, which makes it impossible to add a test case.

* [HTML] Rename and sort attribute name variables

This commit equalizes the attribute_name_... variables with the recent
change of tag_name... variables.

* [HTML] case insensitive doctype matching and break

Fixes #2236
Supersedes  #2237

This commit applies case-insensitive tag names and the tag-name break
fixes to doctype tags as asked by the author of PR #2237 to avoid merge
conflicts.

* [HTML] Cleanup style close tag lookahead

Related to #2321, #2322

Other than script tags, commented out style tag content isn't 
highlighted (see added syntax test). Therefore we don't need to care
about it in the escape lookahead.